pyproject.toml
[tool.poetry]
name = "testbed"
version = "0.1.0"
description = ""
authors = ["Jost Rossel <jrossel@mail.upb.de>"]

[tool.poetry.dependencies]
python = "^3.8, <3.9"
pypandoc = "^1.6.3"
bs4 = "^0.0.1"
"ruamel.yaml" = "^0.17.10"
pywin32 = { version = "^301", optional = true }  # not really optional, but allows dep install on linux
Pillow = "^8.3.1"
xmlschema = "^1.7.1"
requests = "^2.26.0"
scikit-image = "^0.18.3"
numpy = "^1.21.2"
matplotlib = "^3.4.3"
numpy-stl = "^2.16.3"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
isort = "^5.8.0"
pylint = "^2.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.pylint.messages_control]
disable = "C0330, C0326, E1136, too-many-arguments"

[tool.pylint.format]
max-line-length = "100"
